+++ packages/core/lib/match.js
@@ -1,17 +1,20 @@
+const _ = require('lodash')
 const globToRegex = require('glob-to-regexp')
 const { assertConfigSetters, invokeConfigSetters } = require('./configSetters')
 const { invokePreHooks, invokePostHooks } = require('./hooks')
 
 module.exports = match
 
+const regexify = glob => globToRegex(glob, { extended: true })
+const stripArrayConditionally = array => array.length === 1 ? array[0] : array
+const toArray = thing => Array.isArray(thing) ? thing : [ thing ]
+
 /**
  * State on which files to apply the loader blocks passed in this call. Works
  * like `group()`, but adds the file matching information to the context.
  *
  * @param {string|RegExp|Function|Array} test   A glob like `*.css` or `{*.js, *.jsx}` or something else to use as `loader.test`.
- * @param {object} [options]                    Optional advanced matching options.
- * @param {string|Function|RegExp|Array|object} [options.include]
- * @param {string|Function|RegExp|Array|object} [options.exclude]
+ * @param {object} [options]                    Rule options. See https://webpack.js.org/configuration/module/
  * @param {Function[]} configSetters            Array of functions as returned by webpack blocks.
  * @return {Function}
  */
@@ -23,13 +26,21 @@ function match (test, options, configSetters) {
 
   assertConfigSetters(configSetters)
 
-  const match = { test: createFileTypeMatcher(test) }
-
-  if (options.exclude) {
-    match.exclude = options.exclude
+  if (options.test) {
+    throw new Error(`match(): Setting 'test' in options is not supported; use the argument instead.`)
   }
-  if (options.include) {
-    match.include = options.include
+
+  const { inclusions, exclusions } = splitPatterns(toArray(test))
+  const match = Object.assign({}, options, {
+    // The `.test` is usually just one pattern, so stripping the array feels more natural
+    test: stripArrayConditionally(normalizeMatchers(inclusions))
+  })
+
+  if (exclusions.length > 0) {
+    match.exclude = _.concat(
+      match.exclude ? toArray(match.exclude) : [],
+      normalizeMatchers(exclusions)
+    )
   }
 
   const groupBlock = context => config => invokeConfigSetters(configSetters, deriveContextWithMatch(context, match), config)
@@ -40,15 +51,21 @@ function match (test, options, configSetters) {
   })
 }
 
-const regexify = glob => globToRegex(glob, { extended: true })
-
-function createFileTypeMatcher (test) {
-  if (typeof test === 'string') {
-    return regexify(test)
-  } else if (Array.isArray(test) && test.every(item => typeof item === 'string')) {
-    return test.map(item => regexify(item))
-  } else {
+function normalizeMatchers (fileMatchTests) {
+  return fileMatchTests.map(test => {
+    if (typeof test === 'string') {
+      return regexify(test)
+    }
     return test
+  })
+}
+
+function splitPatterns (patterns) {
+  const isNegation = pattern => typeof pattern === 'string' && pattern.startsWith('!')
+  const stripLeadingExclam = pattern => pattern.startsWith('!') ? pattern.substr(1) : pattern
+  return {
+    inclusions: patterns.filter(pattern => !isNegation(pattern)),
+    exclusions: patterns.filter(pattern => isNegation(pattern)).map(stripLeadingExclam)
   }
 }