+++ packages/core/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next version
 
 - Added `when()` ([#242](https://github.com/andywer/webpack-blocks/issues/242))
+- Support negations in `match()` to set `exclude` ([#252](https://github.com/andywer/webpack-blocks/issues/252))
+- Change `match()` to pass all [webpack rule options](https://webpack.js.org/configuration/module/) to context. ([#250](https://github.com/andywer/webpack-blocks/pull/250))
 
 ## 1.0.0-beta.2